Excercise -> Exercise. Renamed files, typo fixed.
I noticed that I had been consistently writing about exCercise when I
meant exercise.

Now the only place typo remains is project history. I could rebase it
all but I don't know if gain justifies effort required and potential
history falsification. Not to mention time this would take.

It has been said that the great scientific disciplines are examples of giants standing on the shoulders of other giants.  It has also been said that the software industry is an example of midgets standing on the toes of other midgets.
      -- Alan Cooper